+++ model/client_pack.go
@@ -5,8 +5,8 @@ type ClientPacks []*ClientPack
 
 // ClientPack represents a client pack model definition.
 type ClientPack struct {
-	ClientID int     `json:"client_id" sql:"index"`
+	ClientID int64   `json:"client_id" sql:"index"`
 	Client   *Client `json:"client,omitempty"`
-	PackID   int     `json:"pack_id" sql:"index"`
+	PackID   int64   `json:"pack_id" sql:"index"`
 	Pack     *Pack   `json:"pack,omitempty"`
 }
