+++ model/team_pack.go
@@ -5,9 +5,9 @@ type TeamPacks []*TeamPack
 
 // TeamPack represents a team pack model definition.
 type TeamPack struct {
-	TeamID int    `json:"team_id" sql:"index"`
+	TeamID int64  `json:"team_id" sql:"index"`
 	Team   *Team  `json:"team,omitempty"`
-	PackID int    `json:"pack_id" sql:"index"`
+	PackID int64  `json:"pack_id" sql:"index"`
 	Pack   *Pack  `json:"pack,omitempty"`
 	Perm   string `json:"perm,omitempty"`
 }
